+++ bot/events/connected.js
@@ -9,6 +9,7 @@ const repo = require(process.cwd()+'/repo');
 
 module.exports = function(bot, db) {
   bot.on('connected', function(data) {
+    bot.isConnected = true;
     bot.log('info', 'BOT', 'Connected to ' + data);
     bot.sendChat("`Initializing...`");
     var initStart = Date.now();