+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The Commands of the bot are its API so here's how it will adhere to Semantic Ver
 
 ## [1.17.1] - 2018-01-05
 ## Fixed
-Fixed [issue 38](https://github.com/FranciscoG/DerpyBot/issues/38) 
+Better error detection and implemented a rate limit queueing system for the chats, hopeully fixing [issue 38](https://github.com/FranciscoG/DerpyBot/issues/38). 
 
 ## [1.17.0] - 2017-12-11
 ### Added